refactor : wallet 관련 레이어 voucher로 이동
